Azimuth: A Shiny App Demonstrating a Query-Reference Mapping Algorithm for Single-Cell Data

What the package does (one paragraph).

Package options

Azimuth uses the following options to control the behavior of the app. Users can provide these as named arguments to AzimuthApp through dots (...), specify these in the config file, or configure these with options.

App options

The following options control app behavior

Azimuth.app.default_adt

ADT to select by default in feature/violin plot

Azimuth.app.default_gene

Gene to select by default in feature/violin plot

Azimuth.app.default_metadata

Default metadata transferred from reference.

Azimuth.app.demodataset

Path to data file (in any Azimuth-supported format) to automatically load when the user clicks a button. The button is only available in the UI if this option is set

Azimuth.app.googlesheet

Google Sheet identifier (appropriate for use with googlesheets4:gs4_get) to write log records. Logging is only enabled if this and other google* options are set

Azimuth.app.googletoken

Path to directory containing Google Authentication token file. Logging is only enabled if this and other google* options are set

Azimuth.app.googletokenemail

Email address corresponding to the Google Authentication token file. Logging is only enabled if this and other google* options are set

Azimuth.app.max_cells

Maximum number of cells allowed to upload

Azimuth.app.metadata_notransfer

Metadata to annotate in reference but not transfer to query

Azimuth.app.mito

Regular expression pattern indicating mitochondrial features in query object

Azimuth.app.plotseed

Seed to shuffle colors for cell types

Azimuth.app.reference

URL or directory path to reference dataset; see LoadReference for more details

Azimuth.app.refuri

URL for publicly available reference dataset, used in the downloadable analysis script in case Azimuth.app.reference points to a directory

Azimuth.app.refdescriptor

Provide (as a string) the html to render the reference description on the welcome page

Azimuth.app.welcomebox

Provide (as a string) the code to render the box on the welcome page (quotes escaped). Example:

box(
  h3(\"Header\"),
  \"body text\",
  a(\"link\", href=\"www.satijalab.org\", target=\"_blank\"),
  width = 12
)

Azimuth.app.homologs

URL or path to file containing the human/mouse homolog table.

Azimuth.app.metatableheatmap

Display the meta.data table as a heatmap rather than in tabular form. defaults to FALSE.

Azimuth.app.overlayedreference

Display the mapped query on top of greyed out reference in the 'Cell Plots' tab. defaults to FALSE

Control options

These options control mapping and analysis behavior

Azimuth.map.ncells

Minimum number of cells required to accept uploaded file defaults to 100

Azimuth.map.ngenes

Minimum number of genes in common with reference to accept uploaded file; defaults to 250

Azimuth.map.nanchors

Minimum number of anchors that must be found to complete mapping. Defaults to 50

Azimuth.map.panchorscolors

Configure the valuebox on the main page corresponding to the values for failure, warning, success for fraction of unique query cells that participate in anchor pairs. Failure corresponds to [0:Azimuth.map.fracanchorscolors[1]), warning to [Azimuth.map.fracanchorscolors[1]:Azimuth.map.fracanchorscolors[2]), and success is >= Azimuth.map.fracanchorscolors[2]. Defaults to c(5, 15)

Azimuth.map.postmapqccolors

Configure the valuebox on the main page corresponding to the values for failure, warning, success for the post mapping cluster based QC metric. Failure corresponds to [0:Azimuth.map.postmapqc[1]), warning to [Azimuth.map.postmapqc[1]:Azimuth.map.postmapqc[2]), and success is >= Azimuth.map.postmapqc[2]. Defaults to c(0.15, 0.25)

Azimuth.map.postmapqcds

Set the amount of query random downsampling to perform before computing the mapping QC metric. Defaults to 5000

Azimuth.map.ntrees

Annoy (approximate nearest neighbor) n.trees parameter Defaults to 20

Azimuth.map.ndims

Number of dimensions to use in FindTransferAnchors and TransferData Defaults to 50

Azimuth.de.mincells

Minimum number of cells per cluster for differential expression; defaults to 15

Azimuth.de.digits

Number of digits to round differential expression table to; defaults to 3

Azimuth.sct.ncells, Azimuth.sct.nfeats

Number of cells and features to use for SCTransform, respectively. Defaults to 2000 for each

External options

The following options are used by external dependencies that have an effect on Azimuth's behavior. Refer to original package documentation for more details

shiny.maxRequestSize

User-configurable; used for controlling the maximum file size of uploaded datasets. Defaults to 500 Mb

DT.options

User-configurable; used for controlling biomarker table outputs. Defaults to setting pageLength to 10

future.globals.maxSize

Non-configurable; used for parallelization. Defaults to Azimuth.app.max_cells * 320000

See also

Author

Maintainer: Paul Hoffman phoffman@nygenome.org (ORCID)

Authors:

Andrew Butler abutler@nygenome.org (ORCID)

Charlotte Darby cdarby@nygenome.org (ORCID)

Yuhan Hao yhao@nygenome.org (ORCID)

Rahul Satija rsatija@nygenome.org (ORCID)

Other contributors:

Jaison Jain jjain@nygenome.org (ORCID) [contributor]

Satija Lab and Collaborators [funder]