Create a Seurat object compatible with Azimuth.

Create a Seurat object compatible with Azimuth.

AzimuthBridgeReference(
  object,
  reference.reduction = "spca",
  bridge.ref.reduction = "ref.spca",
  bridge.query.reduction = "slsi",
  laplacian.reduction = "lap",
  refUMAP = "wnn.umap",
  refAssay = "SCT",
  dims = 1:50,
  plotref = "wnn.umap",
  plot.metadata = NULL,
  ori.index = NULL,
  colormap = NULL,
  assays = c("Bridge", "RNA"),
  metadata = NULL,
  reference.version = "0.0.0",
  verbose = FALSE
)

Arguments

object

Seurat object

refUMAP

Name of UMAP in reference to use for mapping

refAssay

Name of SCTAssay to use in reference

dims

Dimensions to use in reference neighbor finding

plotref

Either the name of the DimReduc in the provided Seurat object to use for the plotting reference or the DimReduc object itself.

plot.metadata

A data.frame of discrete metadata fields for the cells in the plotref.

ori.index

Index of the cells used in mapping in the original object on which UMAP was run. Only need to provide if UMAP was run on different set of cells.

colormap

A list of named and ordered vectors specifying the colors and levels for the metadata. See CreateColorMap for help generating your own.

assays

Assays to retain for transfer

metadata

Metadata to retain for transfer

reference.version

Version of the Azimuth reference

verbose

Display progress/messages

refDR

Name of DimReduc in reference to use for mapping

k.param

Defines k for the k-nearest neighbor algorithm

Value

Returns a Seurat object with AzimuthData stored in the tools slot for use with Azimuth.